[Fix] 토큰 갱신 로직 수정
